Add files via upload
As per the data 39,717 number of Bank Customers. We clean date like removing duplicate and irreverent data, normalize and standardize the format We are using different types of KPI's to analyze the data and find out the verified and non-verified customer based on year and interest. How payment mode is affected in loan section. We believe that data-driven insight and actionable recommendation for stakeholders in the financial industry.